cisco_vip_network_tool/cisco_vip_network_tool/requirements.txt
networkx>=3.2
matplotlib>=3.8
pyyaml>=6.0
scapy>=2.5.0
